Cleaned up components, images, and renamed Routes
