+++ bloom-and-vine/src/App.js
@@ -1,4 +1,4 @@
-import { BrowserRouter, Routes, Route } from "react-router-dom"
+import { Routes, Route } from "react-router-dom"
 import Homepage from './Pages/Homepage'
 import Gallery from './Pages/Gallery'
 import Services from './Pages/Services'
@@ -8,14 +8,14 @@ import Contact from './Pages/Contact';
 function App() {
   return (
     <div className="App">
-      <Routes>
-        <Route exact path='/' element={<Homepage />} />
-        <Route exact path='/gallery' element={<Gallery />} />
-        <Route exact path='/services' element={<Services />} />
-        <Route exact path='/about' element={<About />} />
-        <Route exact path='/contact' element={<Contact />} />
-        <Route path="*" element={<div>error</div>} />
-      </Routes>
+        <Routes>
+            <Route exact path='/Home' element={<Homepage />} />
+            <Route exact path='/Gallery' element={<Gallery />} />
+            <Route exact path='/Services' element={<Services />} />
+            <Route exact path='/About' element={<About />} />
+            <Route exact path='/Contact' element={<Contact />} />
+            <Route path="*" element={<div>error</div>} />
+        </Routes>
     </div>
   );
 }
