+++ bloom-and-vine/src/Components/instagram-posts-carousel/InstagramPostsCarousel.css
@@ -81,5 +81,4 @@
    .backward-posts-button, .forward-posts-button {
         display: none;
     }
-
 }
