+++ bloom-and-vine/src/Components/redirect-images/RedirectDisplay.js
@@ -1,8 +1,8 @@
 import RedirectImage from './components/redirect-image/RedirectImage'
-import GalleryImg from '../../Images/gallery.jpg'
-import ServicesImg from '../../Images/services.jpg'
-import AboutImg from '../../Images/about.jpg'
-import ContactImg from '../../Images/contact.jpg'
+import GalleryImg from '../../Images/galleryRedirect.jpg'
+import ServicesImg from '../../Images/servicesRedirect.jpg'
+import AboutImg from '../../Images/aboutRedirect.jpg'
+import ContactImg from '../../Images/contactRedirect.jpg'
 import './RedirectDisplay.css'
 
 export default function RedirectDisplay() {
