+++ bloom-and-vine/src/Components/redirect-images/components/redirect-image/RedirectImage.js
@@ -1,4 +1,4 @@
-import { BrowserRouter as Router, Route, NavLink } from 'react-router-dom'
+import { NavLink } from 'react-router-dom'
 import './RedirectImage.css'
 
 export default function RedirectImage(props) {
@@ -6,12 +6,12 @@ export default function RedirectImage(props) {
 
     return (
         <div className='image-container'>
-            <div className='image-background' id={redirect.toLowerCase()}>
+            <div className='image-background' id={redirect}>
                 <p className='image-text'>
                     {redirect}
                 </p>
             </div>
-            <NavLink to={`/${redirect.toLowerCase()}`}>
+            <NavLink to={`/${redirect}`}>
                 <img className='image-redirect' src={img} loading='eager'/>
             </NavLink>
         </div>
