+++ bloom-and-vine/src/Pages/Homepage.js
@@ -1,17 +1,11 @@
 import { useState, useEffect } from 'react';
-import Navbar from '../Components/Navbar'
-import Footer from '../Components/Footer'
+import axios from 'axios';
 import ReviewSlider from '../Components/review-slider/ReviewSlider';
 import RedirectImages from '../Components/redirect-images/RedirectDisplay';
 import InstagramShoutout from '../Components/instagram-shoutout/InstagramShoutout';
 import InstagramPostsCarousel from '../Components/instagram-posts-carousel/InstagramPostsCarousel';
-import frontDisplayImg from '../Images/homepage_front_display.jpg'
-import axios from 'axios';
+import homepageFrontDisplay from '../Images/homepageFrontDisplay.jpg'
 import './Homepage.css';
-
-// Put equal spacing between img, reviews, & instagram shoutout (DONE)
-
-// Indication that redirect has been clicked (darkened image) for mobile (done?)
 
 const options = {
     method: 'GET',
@@ -36,10 +30,8 @@ function Homepage() {
 
 	return (
 		<div className="Homepage">
-			<Navbar />
-
 			<div className='front-display'>
-				<img className="front-image" src={frontDisplayImg}/>
+				<img className="front-image" src={homepageFrontDisplay}/>
 				<div className='front-text-container'>
 					<p className='front-text'>Bloom & Vine</p>
 				</div>
@@ -60,8 +52,6 @@ function Homepage() {
 			<div className='instagram-posts-display'>
 				{postData && <InstagramPostsCarousel data={postData} />}
 			</div>
-		
-			<Footer />
 		</div>
   	);
 }
